login view common
